requirements.txt
Package           Version
----------------- -------
async-timeout     4.0.3
asyncpg           0.29.0
greenlet          3.0.3
pip               21.3.1
psycopg2          2.9.9
pyaes             1.6.1
Pyrogram          2.0.106
PySocks           1.7.1
python-dotenv     1.0.1
setuptools        60.2.0
SQLAlchemy        2.0.29
typing_extensions 4.11.0
wheel             0.37.1
